requirements.txt
streamlit==1.40.1
requests==2.31.0
Pillow==11.0.0
beautifulsoup4==4.12.2
groq==0.4.2
